+++ src/scriptlets/set-constant.js
@@ -285,6 +285,30 @@ export function setConstant(source, property, value, stack = '', valueWrapper = 
                 if (prevSetter !== undefined) {
                     prevSetter(a);
                 }
+                // Set a proxy trap to observe changes
+                // This is a partial fix and only works with a single scriptlet,
+                // a full fix requires synchronisation between the scriptlets
+                // TODO: add proper fix when synchronisation between scriptlets is added
+                // https://github.com/AdguardTeam/Scriptlets/issues/330
+                if (a instanceof Object) {
+                    // Get properties which should be checked and remove first one
+                    // because it's current object
+                    const propertiesToCheck = property.split('.').slice(1);
+                    a = new Proxy(a, {
+                        get: (target, propertyKey, val) => {
+                            // Check if object contains required property, if so
+                            // check if current value is equal to constantValue, if not, set it to constantValue
+                            propertiesToCheck.reduce((object, currentProp, index, array) => {
+                                const currentObj = object?.[currentProp];
+                                if (currentObj && index === array.length - 1 && currentObj !== constantValue) {
+                                    object[currentProp] = constantValue;
+                                }
+                                return currentObj || object;
+                            }, target);
+                            return Reflect.get(target, propertyKey, val);
+                        },
+                    });
+                }
                 handler.set(a);
             },
         });